moved HTML parser to parsers package
